[daemon] Be less aggressive to store the primary clipboard
The primary clipboard was saved very, very aggressively. This showed a
real slow down when doing selections. Now it is delayed inside a timeout
at a quarter second interval and will save for real once the selection
is done (no mouse button hold or shift button on keyboard).
If the selection is lost, we avoid storing the cached value again by
keeping a state with an internal boolean value.