[daemon] Fix bad assumption that I could free/copy the selection
The code handling the get-selection callback was changed to free the
selection and than copy the cached selection into it. This was baaad as
it made further calls to the get-selection callback when freeing the
selection which ended in a crash. The code has been switched back to the
heavier version of calling the set function.