daemon: dispatch primary clipboard restoration
The GtkEntryCompletion widget is continuously removing ownership over
the clipboard, quickly after that setting it back.

The code assumed that when the ownership was removed, it could safely
restore the cached value. Wrong. This causes the widget to "erase" its
current selection and typing continues at the end of the string (as seen
in GNOME bug 333320).

Now restoring works through a dispatched GSource, which enables to
remove this GSource if ownership is set right back again.