New feature "paste-on-activate"
There is a new property in the ClipmanMenu class "paste-on-activate" that
is bound to the xfconf property "/tweaks/paste-on-activate". This new
property gives the possibility to paste a content that has been selected
in the history.

Currently the default value is off, there are two possible values to
paste the content with Ctrl+V or Shift+Insert.