Make sure to overwrite binary if it already exists
If you locally compile and install multiple times, before this commit
you would have to delete the symbolic link manually because ln -s
would not overwrite it, hence resulting in make install failing.
The -f option fixes this and should also work on systems without ln,
as it falls back to cp.