plugin: "Show full history..." menuitem
Since we don't show the full history in the plugin anymore let's add a
hint that we don't and make it a launcher for xfce4-clipman-history. We
only check if the *maximum* number of items in the history is greater
than the one displayed in the panel, not the *actual* number of items.