Fix Gtk-CRITICAL when showing main menu
`gtk_clipboard_wait_for_text()` is currently used to request clipboard
text when populating the main menu. This synchronous request causes the
event that triggers menu popup to be dispatched, resulting in a
Gtk-CRITICAL when showing the menu. So let's use
`gtk_clipboard_request_text()` instead, with the necessary adaptations.