settings: Fix 'skip-action-on-key-down' sync and improve strings
The radio button can't be synchronized by a simple binding with a
boolean, it would have to be connected to a signal handler acting
alternately on one button and the other.

But it's better to use a combo box instead. It can be synchronized
directly, and a single label with several choices is easier to
understand in this case.

Helps: #51, #52