history: Refactor item_to_restore into image_to_restore
{set,get}_item_to_restore() are actually only intended to handle the
special case of images, as their documentation clearly states, because
otherwise string comparisons can be made. It is therefore unnecessary to
maintain a generic item_to_restore, which complicates the code and is
error prone.

The use of the go-next-symbolic icon in the menu is removed, as it
doesn't correspond to anything precise and misleads the user as to what
is contained in his clipboards.

Finally, images are now correctly reordered in the history after
selection from the menu.