Move default clipboard restoration from manager to collector
In a simplified form, where only text is restored, similar to what is
done for the primary clipboard. This ensures the same behavior, whether
Clipman's clipboard manager is used or not (#28). It also allows all
restoration operations to be carried out in a single place, for greater
control, especially when the history is cleared (#4).

Fixes: #4, #28