settings: Better fit with the way GApplication works
This leaves the application's uniqueness to be managed internally, and
uses a "command-line" signal handler to manage the return code from the
primary instance.

This ensures that the primary instance has been activated before exiting
from the remote instance and removes the warning to use
g_application_run(). Error handling is improved by the way.

Finally, this adds the settings dialog as app window and uses shutdown()
to free data instead of using gtk_dialog_run().