Replace deprecated gtk_menu_popup() when possible
It is not possible for GtkStatusIcon, which is itself deprecated, X11
only and therefore to be replaced with StatusNotifierItem in the future.
Clipboard content must now be retrieved synchronously in the
button-press-event signal handler and before showing the menu.
This fixes various menu positioning problems when autohide is enabled
and is part of the preparation for porting the plugin to Wayland.

Fixes: #86