Use X11 clipboard manager from libxfce4ui if available
This will avoid inconsistencies in behavior depending on whether
xfsettingsd is running or not, and any bugs can be fixed in one place.
This has become possible in particular since the code that wasn't purely
X11 above was commented out and transferred to ClipmanCollector in
eab4beac and a831330a.

Fixes: #107
Related: xfce/libxfce4ui!115